requirements.txt
mysql-connector-python
langchain
langchain_core
langchain-ollama
flask
flask-cors
matplotlib
pandas
python-dotenv